Refactor: Organize code formatting
